Refactored summary banners
