+++ components/clinical/shared/summary/src/index.tsx
@@ -1,1 +1,2 @@
 export * from './atoms'
+export * from './organisms'
