--- old/summary/CHANGELOG.md
@@ -1,187 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [0.0.24](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.22...@ltht-react/summary@0.0.24) (2020-11-19)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.23](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.22...@ltht-react/summary@0.0.23) (2020-10-29)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.22](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.21...@ltht-react/summary@0.0.22) (2020-09-10)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-
-## [0.0.21](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.20...@ltht-react/summary@0.0.21) (2020-09-01)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.20](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.19...@ltht-react/summary@0.0.20) (2020-08-21)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.19](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.18...@ltht-react/summary@0.0.19) (2020-08-19)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.18](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.17...@ltht-react/summary@0.0.18) (2020-08-19)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.17](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.16...@ltht-react/summary@0.0.17) (2020-08-18)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.16](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.15...@ltht-react/summary@0.0.16) (2020-08-11)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.15](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.14...@ltht-react/summary@0.0.15) (2020-08-10)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.14](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.13...@ltht-react/summary@0.0.14) (2020-08-05)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.13](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.12...@ltht-react/summary@0.0.13) (2020-08-05)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.12](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.11...@ltht-react/summary@0.0.12) (2020-08-05)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.10](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.9...@ltht-react/summary@0.0.10) (2020-08-04)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-
-## [0.0.9](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.8...@ltht-react/summary@0.0.9) (2020-05-06)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.8](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.7...@ltht-react/summary@0.0.8) (2020-05-01)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.7](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.6...@ltht-react/summary@0.0.7) (2020-05-01)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.6](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.5...@ltht-react/summary@0.0.6) (2020-04-29)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.5](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.4...@ltht-react/summary@0.0.5) (2020-04-28)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.4](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.3...@ltht-react/summary@0.0.4) (2020-04-14)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.3](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.2...@ltht-react/summary@0.0.3) (2020-04-09)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-## [0.0.2](https://github.com/ltht-epr/ltht-react/compare/@ltht-react/summary@0.0.1...@ltht-react/summary@0.0.2) (2020-04-08)
-
-**Note:** Version bump only for package @ltht-react/summary
-
-
-
-
-
-
-## 0.0.1 (2020-04-03)
-
-**Note:** Version bump only for package @ltht-react/summary
