--- old/summary/README.md
@@ -1,17 +1,0 @@
-
-# Summary
-
-<!-- STORY -->
-
-### Import
-
-```js
-import Summary from '@ltht-react/summary'
-```
-
-### Usage
-
-```jsx
-<Summary />
-```
-
