--- old/summary/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@ltht-react/summary",
-  "version": "0.0.24",
-  "description": "ltht-react clinical Summary component.",
-  "author": "LTHT",
-  "homepage": "",
-  "license": "MIT",
-  "main": "lib/index.js",
-  "typings": "lib/index.d.ts",
-  "files": [
-    "lib"
-  ],
-  "directories": {
-    "lib": "lib"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "clean": "rimraf lib",
-    "build": "yarn build:lib",
-    "watch": "tsc -w",
-    "build:lib": "tsc",
-    "type:check": "tsc --noEmit",
-    "lint": "prettylint 'src/*.{ts,tsx}'"
-  },
-  "dependencies": {
-    "@emotion/core": "^10.0.27",
-    "@emotion/styled": "^10.0.27",
-    "@ltht-react/banner": "^0.0.18",
-    "@ltht-react/icon": "^0.1.28",
-    "@ltht-react/list": "^0.1.26",
-    "@ltht-react/styles": "^0.1.20",
-    "@ltht-react/types": "^0.0.27",
-    "@ltht-react/utils": "^0.1.20",
-    "emotion": "^10.0.27",
-    "react": "^16.12.0"
-  },
-  "gitHead": "4d364556bb11e294d86697c2dac66742f7aabc7f"
-}
