--- old/summary/src/index.tsx
@@ -1,2 +1,0 @@
-export * from './atoms'
-export * from './organisms'
